+++ builder.go
@@ -15,12 +15,9 @@ func undefinedError(v variable.Any) error {
 }
 
 // isBuilderOf makes a static assertion that B meats
-type isBuilderOf[T any, B builderOf[T]] struct{}
-
-// builderOf is an interface and type constriant common to all builders that
-// produce a value of type T.
-type builderOf[T any] interface {
+type isBuilderOf[T any, B interface {
+	SeeAlso(input Input, options ...SeeAlsoOption) B
 	Required(options ...RequiredOption) Required[T]
 	Optional(options ...OptionalOption) Optional[T]
 	Deprecated(options ...DeprecatedOption) Deprecated[T]
-}
+}] struct{}