+++ internal/mode/usage/markdown/spec.go
@@ -41,6 +41,7 @@ func (r *specRenderer) Render() {
 	}
 
 	r.renderUnimportantDocumentation()
+	r.renderRelationships()
 }
 
 // VisitNumeric renders the primary requirement for spec that uses the "numeric"