+++ relationshipexample_test.go
@@ -1,0 +1,18 @@
+package ferrite_test
+
+import (
+	"github.com/dogmatiq/ferrite"
+)
+
+func Example_seeAlso() {
+	defer example()()
+
+	verbose := ferrite.
+		Bool("VERBOSE", "enable verbose logging").
+		Optional()
+
+	ferrite.
+		Bool("DEBUG", "enable or disable debugging features").
+		SeeAlso(verbose).
+		Optional()
+}